requirements.txt
Wand
twython
wget
Pillow
PyPDF2
selenium
beautifulsoup4
lxml
python-slugify